+++ examples/05.submitOrder.js
@@ -10,7 +10,7 @@ Check README.md for more details.
 const sw = require('@rhino.fi/starkware-crypto')
 const getWeb3 = require('./helpers/getWeb3')
 
-const DVF = require('../src/dvf')
+const RhinofiClientFactory = require('../src')
 const envVars = require('./helpers/loadFromEnvOrConfig')(
   process.env.CONFIG_FILE_NAME
 )
@@ -18,21 +18,27 @@ const logExampleResult = require('./helpers/logExampleResult')(__filename)
 
 const ethPrivKey = envVars.ETH_PRIVATE_KEY
 // NOTE: you can also generate a new key using:`
-// const starkPrivKey = dvf.stark.createPrivateKey()
+// const starkPrivKey = rhinofi.stark.createPrivateKey()
 const starkPrivKey = envVars.STARK_PRIVATE_KEY
 const rpcUrl = envVars.RPC_URL
 
 const { web3, provider } = getWeb3(ethPrivKey, rpcUrl)
 
-const dvfConfig = {
+const rhinofiConfig = {
   api: envVars.API_URL,
   dataApi: envVars.DATA_API_URL,
-  useAuthHeader: true
+  useAuthHeader: true,
+  wallet: {
+    type: 'tradingKey',
+    meta: {
+      starkPrivateKey: starkPrivKey
+    }
+  }
   // Add more variables to override default values
 }
 
 ;(async () => {
-  const dvf = await DVF(web3, dvfConfig)
+  const rhinofi = await RhinofiClientFactory(web3, rhinofiConfig)
 
   const getPriceFromOrderBook = require('./helpers/getPriceFromOrderBook')
 
@@ -43,11 +49,11 @@ const dvfConfig = {
   const feeRate = ''
 
   // Gets the price from the order book api and cuts 5% to make sure the order will be settled
-  const tickersData = await dvf.getTickers('ETH:USDT');
+  const tickersData = await rhinofi.getTickers('ETH:USDT');
   const orderBookPrice = getPriceFromOrderBook(tickersData);
   const price = orderBookPrice - orderBookPrice * 0.05;
 
-  const submitOrderResponse = await dvf.submitOrder({
+  const submitOrderResponse = await rhinofi.submitOrder({
     symbol,
     amount,
     price,