+++ examples/13.withdraw.js
@@ -10,7 +10,7 @@ Check README.md for more details.
 const sw = require('@rhino.fi/starkware-crypto')
 const getWeb3 = require('./helpers/getWeb3')
 
-const DVF = require('../src/dvf')
+const RhinofiClientFactory = require('../src')
 const envVars = require('./helpers/loadFromEnvOrConfig')(
   process.env.CONFIG_FILE_NAME
 )
@@ -18,13 +18,13 @@ const logExampleResult = require('./helpers/logExampleResult')(__filename)
 
 const ethPrivKey = envVars.ETH_PRIVATE_KEY
 // NOTE: you can also generate a new key using:`
-// const starkPrivKey = dvf.stark.createPrivateKey()
+// const starkPrivKey = rhinofi.stark.createPrivateKey()
 const starkPrivKey = envVars.STARK_PRIVATE_KEY
 const rpcUrl = envVars.RPC_URL
 
 const { web3, provider } = getWeb3(ethPrivKey, rpcUrl)
 
-const dvfConfig = {
+const rhinofiConfig = {
   api: envVars.API_URL,
   dataApi: envVars.DATA_API_URL,
   useAuthHeader: true,
@@ -38,15 +38,16 @@ const dvfConfig = {
 }
 
 ;(async () => {
-  const dvf = await DVF(web3, dvfConfig)
+  const rhinofi = await RhinofiClientFactory(web3, rhinofiConfig)
 
   const token = 'ETH'
   const amount = 0.1
 
-  const withdrawalResponse = await dvf.transferAndWithdraw({
-    recipientEthAddress: dvf.get('account'),
+  // NOTE: withdraw method as been deprecated
+  const withdrawalResponse = await rhinofi.transferAndWithdraw({
+    recipientEthAddress: rhinofi.get('account'),
     token,
-    amount
+    amount,
   })
 
   logExampleResult(withdrawalResponse)