+++ examples/src/ledgerDeposit.js
@@ -2,13 +2,13 @@ const path = `44'/60'/0'/0'/0`
 const token = 'ETH'
 const amount = 0.70
 
-const starkDepositData = await dvf.stark.ledger.createDepositData(
+const starkDepositData = await rhinofi.stark.ledger.createDepositData(
   path,
   token,
   amount
 )
 
-const depositResponse = await dvf.ledger.deposit(
+const depositResponse = await rhinofi.ledger.deposit(
   token,
   amount,
   starkDepositData