+++ examples/src/register.js
@@ -1,5 +1,5 @@
-const keyPair = await dvf.stark.createKeyPair(starkPrivKey)
+const keyPair = await rhinofi.stark.createKeyPair(starkPrivKey)
 
-const registerResponse = await dvf.register(keyPair.starkPublicKey)
+const registerResponse = await rhinofi.register(keyPair.starkPublicKey)
 
 logExampleResult(registerResponse)