+++ examples/src/transfer.js
@@ -1,0 +1,7 @@
+const transferResponse = await rhinofi.transfer({
+  recipientEthAddress: rhinofi.config.DVF.deversifiAddress,
+  token: 'USDT',
+  amount: 5
+})
+
+logExampleResult(transferResponse)