+++ examples/src/withdraw.js
@@ -1,10 +1,11 @@
 const token = 'ETH'
 const amount = 0.1
 
-const withdrawalResponse = await dvf.withdraw(
+// NOTE: withdraw method as been deprecated
+const withdrawalResponse = await rhinofi.transferAndWithdraw({
+  recipientEthAddress: rhinofi.get('account'),
   token,
   amount,
-  starkPrivKey
-)
+})
 
 logExampleResult(withdrawalResponse)