+++ nix/packages/append-dev-to-node-package-version.nix
@@ -1,0 +1,16 @@
+{
+  writers,
+  bash,
+  lib,
+  moreutils,
+  jq,
+  name ? "append-dev-to-node-package-version"
+}:
+writers.writeBashBin name ''
+  set -ueo pipefail
+
+  ${lib.getExe jq} \
+    '.version = .version + "-dev"' \
+    package.json \
+    | ${moreutils}/bin/sponge package.json
+''