+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 5.0.0 - 2023-01-07
+### Changed
+- rename from dvf-client-js to @rhino.fi/client-js
+- rename entry point from src/dvf to src/index
+- remove references to dvf from README and examples
+
 ## 4.2.0 - 2023-01-06
 ### Changed
 - Add new methods